Add .pre-commit-hooks.yaml
Pre-commit (https://pre-commit.com) is a tool used to manage pre-commit
hooks. It can download tools from github and runs themn isolated
environments. It uses a small configuration file: .pre-commit-hooks.yaml
placed in the root folder of the repositories it has been instructed to
fetch to discover what tools can it provides.

With this change, I add a `.pre-commit-hooks.yaml` file which lists
`cmakelint` as the only tool.