Remove conflicting/duplicate scripts
Specifying both the script `bin/cmakelint` and the entry-point
`cmakelint` is redundant and, more seriously, results in a conflict when
installing the package:

    $ python -m build --wheel --no-isolation
    ...
    Successfully built cmakelint-1.4.2-py3-none-any.whl
    $ python -m installer --destdir=tmp_install dist/*.whl
    ...
    FileExistsError: File already exists: tmp_install/usr/bin/cmakelint

As such, remove the `scripts` declaration and only keep the
`entry_point` one.

Leave the actual script `bin/cmakelint` for users who want to run the
program from the repository without installing the Python package.